server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "fastapi-plugins>=0.13.1",
    "sqlmodel>=0.0.22",
    "strawberry-graphql[fastapi]>=0.256.1",
    "alembic>=1.14.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "pyjwt>=2.10.1",
]
